Refactored DETECTORS, LOCATIONS, USERS, and experience helper functions into separate files
