+++ imports/api/Testing/LOCATIONS.js
@@ -1,0 +1,26 @@
+export let LOCATIONS = {
+  'park': {
+    lat: 42.056838,
+    lng: -87.675940
+  },
+  'lakefill': {
+    lat: 42.054902,
+    lng: -87.670197
+  },
+  'burgers': {
+    lat: 42.046131,
+    lng: -87.681559
+  },
+  'grocery': {
+    lat: 42.047621,
+    lng: -87.679488
+  },
+  'sushi': { // sashimi sashimi near whole foods
+    lat: 42.048068,
+    lng: -87.6811262
+  },
+  'grocery2': {
+    lat: 42.039818,
+    lng: -87.680088
+  }
+};
