+++ imports/api/Testing/USERS.js
@@ -1,0 +1,77 @@
+export let USERS = {
+  garrett: {
+    username: 'garrett',
+    email: 'garret@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Garrett',
+      lastName: 'Hedman',
+      staticAffordances: {
+        mechanismRich: true
+      }
+    }
+  },
+  garretts_brother: {
+    username: 'garretts_brother',
+    email: 'garretts_brother@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Barrett', // theres an inside joke to this one
+      lastName: 'Hedman'
+    }
+  },
+  meg: {
+    username: 'meg',
+    email: 'meg@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Meg',
+      lastName: 'Grasse'
+    }
+  },
+  megs_sister: {
+    username: 'megs_sister',
+    email: 'megs_sister@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Sister of Meg',
+      lastName: 'Grasse'
+    }
+  },
+  andrew: {
+    username: 'andrew',
+    email: 'andrew@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Andrew',
+      lastName: 'Finke'
+    }
+  },
+  josh: {
+    username: 'josh',
+    email: 'josh@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Josh',
+      lastName: 'Shi'
+    }
+  },
+  nagy: {
+    username: 'nagy',
+    email: 'nagy@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Nagy',
+      lastName: 'Hakim'
+    }
+  },
+  bonnie: {
+    username: 'bonnie',
+    email: 'bonnie@email.com',
+    password: 'password',
+    profile: {
+      firstName: 'Bonnie',
+      lastName: 'Ishiguro'
+    }
+  }
+};
